+++ src/App.css
@@ -32,7 +32,6 @@
   gap: 20px;
 }
 
-
 .mission-section {
   display: flex;
   justify-content: center;
@@ -59,4 +58,4 @@
 
 .col2 {
   text-align: justify;
-
+}
