--- src/Links.js/Missions.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Missions = () => (
-  <div className="missions">
-    <h2>This is a Missions page</h2>
-    <p>All content about the Missions goes here</p>
-  </div>
-);
-export default Missions;
